refactor: decompose session screen into components and hooks (BLD-329)
Extract from app/session/[id].tsx (2201 → 537 lines):

Components (components/session/):
- SetRow.tsx: set row with RPE/half-step UI
- ExerciseGroupCard.tsx: exercise card with sets table
- ExerciseDetailDrawer.tsx: exercise detail bottom drawer
- SetTypeSheet.tsx: set type picker overlay
- types.ts: shared types (SetWithMeta, ExerciseGroup)

Hooks (hooks/):
- useSessionData.ts: session loading, initialization, data state
- useExerciseManagement.ts: swap, delete, add exercise handlers
- useSetTypeActions.ts: set type cycling, PR detection
- useRestTimer.ts: rest timer state, haptics, audio

All 105 test suites / 1609 tests pass. Zero tsc errors.
