+++ __tests__/acceptance/settings-card-notes.acceptance.test.tsx
@@ -28,10 +28,10 @@ describe("Settings profile card layout (BLD-258, GitHub #125)", () => {
 });
 
 describe("Session notes/delete button touch targets (BLD-258, GitHub #126)", () => {
-  const sessionSource = fs.readFileSync(
-    path.resolve(__dirname, "../../app/session/[id].tsx"),
-    "utf-8"
-  );
+  const sessionSource = [
+    fs.readFileSync(path.resolve(__dirname, "../../components/session/ExerciseGroupCard.tsx"), "utf-8"),
+    fs.readFileSync(path.resolve(__dirname, "../../components/session/SetRow.tsx"), "utf-8"),
+  ].join("\n");
 
   it("action buttons have hitSlop for 48dp touch targets", () => {
     // All three action buttons (checkbox, notes, delete) should have hitSlop
