+++ components/session/types.ts
@@ -1,0 +1,22 @@
+import type { WorkoutSet, TrainingMode } from "../../lib/types";
+
+export type SetWithMeta = WorkoutSet & {
+  exercise_name?: string;
+  exercise_deleted?: boolean;
+  previous?: string;
+};
+
+export type ExerciseGroup = {
+  exercise_id: string;
+  name: string;
+  sets: SetWithMeta[];
+  link_id: string | null;
+  training_modes: TrainingMode[];
+  is_voltra: boolean;
+};
+
+export const RPE_CHIPS = [6, 7, 8, 9, 10] as const;
+
+export const RPE_LABELS: Record<number, string> = {
+  6: "Easy", 7: "Easy", 8: "Mod", 9: "Hard", 10: "Max",
+};
